Remove unused code. Reduce class nesting.
