+++ src/Microsoft.Net.Http.Server/NativeInterop/RequestQueue.cs
@@ -28,7 +28,7 @@ namespace Microsoft.Net.Http.Server
     internal class RequestQueue
     {
         private static readonly int BindingInfoSize =
-            Marshal.SizeOf<UnsafeNclNativeMethods.HttpApi.HTTP_BINDING_INFO>();
+            Marshal.SizeOf<HttpApi.HTTP_BINDING_INFO>();
 
         private readonly UrlGroup _urlGroup;
         private readonly ILogger _logger;
@@ -39,8 +39,8 @@ namespace Microsoft.Net.Http.Server
             _logger = logger;
 
             HttpRequestQueueV2Handle requestQueueHandle = null;
-            var statusCode = UnsafeNclNativeMethods.SafeNetHandles.HttpCreateRequestQueue(
-                    UnsafeNclNativeMethods.HttpApi.Version, null, null, 0, out requestQueueHandle);
+            var statusCode = HttpApi.HttpCreateRequestQueue(
+                    HttpApi.Version, null, null, 0, out requestQueueHandle);
 
             if (statusCode != UnsafeNclNativeMethods.ErrorCodes.ERROR_SUCCESS)
             {
@@ -69,13 +69,13 @@ namespace Microsoft.Net.Http.Server
             // Set the association between request queue and url group. After this, requests for registered urls will 
             // get delivered to this request queue.
 
-            var info = new UnsafeNclNativeMethods.HttpApi.HTTP_BINDING_INFO();
-            info.Flags = UnsafeNclNativeMethods.HttpApi.HTTP_FLAGS.HTTP_PROPERTY_FLAG_PRESENT;
+            var info = new HttpApi.HTTP_BINDING_INFO();
+            info.Flags = HttpApi.HTTP_FLAGS.HTTP_PROPERTY_FLAG_PRESENT;
             info.RequestQueueHandle = Handle.DangerousGetHandle();
 
             var infoptr = new IntPtr(&info);
 
-            _urlGroup.SetProperty(UnsafeNclNativeMethods.HttpApi.HTTP_SERVER_PROPERTY.HttpServerBindingProperty,
+            _urlGroup.SetProperty(HttpApi.HTTP_SERVER_PROPERTY.HttpServerBindingProperty,
                 infoptr, (uint)BindingInfoSize);
         }
 
@@ -87,21 +87,21 @@ namespace Microsoft.Net.Http.Server
             // is fine since http.sys allows to set HttpServerBindingProperty multiple times for valid 
             // Url groups.
 
-            var info = new UnsafeNclNativeMethods.HttpApi.HTTP_BINDING_INFO();
-            info.Flags = UnsafeNclNativeMethods.HttpApi.HTTP_FLAGS.NONE;
+            var info = new HttpApi.HTTP_BINDING_INFO();
+            info.Flags = HttpApi.HTTP_FLAGS.NONE;
             info.RequestQueueHandle = IntPtr.Zero;
 
             var infoptr = new IntPtr(&info);
 
-            _urlGroup.SetProperty(UnsafeNclNativeMethods.HttpApi.HTTP_SERVER_PROPERTY.HttpServerBindingProperty,
+            _urlGroup.SetProperty(HttpApi.HTTP_SERVER_PROPERTY.HttpServerBindingProperty,
                 infoptr, (uint)BindingInfoSize, throwOnError: false);
         }
 
         // The listener must be active for this to work.
         internal unsafe void SetLengthLimit(long length)
         {
-            var result = UnsafeNclNativeMethods.HttpApi.HttpSetRequestQueueProperty(Handle,
-                UnsafeNclNativeMethods.HttpApi.HTTP_SERVER_PROPERTY.HttpServerQueueLengthProperty,
+            var result = HttpApi.HttpSetRequestQueueProperty(Handle,
+                HttpApi.HTTP_SERVER_PROPERTY.HttpServerQueueLengthProperty,
                 new IntPtr((void*)&length), (uint)Marshal.SizeOf<long>(), 0, IntPtr.Zero);
 
             if (result != 0)
